requirements.txt
python-dotenv==0.14.0
requests==2.23.0
instabot==0.117.0
Pillow==7.2.0